concatenates two strings

+++ malloc_free/2-str_concat.c
@@ -11,7 +11,7 @@ char *str_concat(char *s1, char *s2)
 {
 	char *result;
 
-	size_t len1, len2, total_len;
+	int len1, len2, total_len;
 
 	if (s1 == NULL)
 	{
